+++ src/api_test.ts
@@ -7,7 +7,6 @@ import {
   mf,
 } from "./dev_deps.ts";
 import { SlackAPI } from "./mod.ts";
-import { serializeData } from "./base-client.ts";
 import { HttpError } from "./deps.ts";
 
 Deno.test("SlackAPI class", async (t) => {
@@ -367,41 +366,6 @@ Deno.test("SlackAPI class", async (t) => {
   );
 
   mf.uninstall();
-});
-
-Deno.test("serializeData helper function", async (t) => {
-  await t.step(
-    "should serialize string values as strings and return a URLSearchParams object",
-    () => {
-      assertEquals(
-        serializeData({ "batman": "robin" }).toString(),
-        "batman=robin",
-      );
-    },
-  );
-  await t.step(
-    "should serialize non-string values as JSON-encoded strings and return a URLSearchParams object",
-    () => {
-      assertEquals(
-        serializeData({ "hockey": { "good": true, "awesome": "yes" } })
-          .toString(),
-        "hockey=%7B%22good%22%3Atrue%2C%22awesome%22%3A%22yes%22%7D",
-      );
-    },
-  );
-  await t.step(
-    "should not serialize undefined values",
-    () => {
-      assertEquals(
-        serializeData({
-          "hockey": { "good": true, "awesome": "yes" },
-          "baseball": undefined,
-        })
-          .toString(),
-        "hockey=%7B%22good%22%3Atrue%2C%22awesome%22%3A%22yes%22%7D",
-      );
-    },
-  );
 });
 
 Deno.test("SlackApi.setSlackApiUrl()", async (t) => {
